Week3/Task1/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
markdown==3.5.2
